Rename Websocket -> ElementsListOld view
Fix #252
